Agrego modal de confirmación al eliminar una tarea

+++ src/app/components/list-task/list-task.component.ts
@@ -8,6 +8,8 @@ import { TaskService } from '../../services/task.service';
 import { Observable } from 'rxjs';
 import { FormsModule } from '@angular/forms';
 import { StatusTaskDirective } from '../../directives/status-task.directive';
+import { MatDialogModule, MatDialog } from '@angular/material/dialog';
+import { ModalComponent } from '../modal/modal.component';
 
 @Component({
     selector: 'app-list-task',
@@ -18,7 +20,9 @@ import { StatusTaskDirective } from '../../directives/status-task.directive';
         MatIconModule,
         MatCheckboxModule,
         FormsModule,
-        StatusTaskDirective
+        StatusTaskDirective,
+        MatDialogModule,
+        ModalComponent
     ],
     templateUrl: './list-task.component.html',
     styleUrl: './list-task.component.scss'
@@ -27,12 +31,21 @@ export class ListTaskComponent {
     displayedColumns: string[] = ['completed', 'position', 'task', 'priority', 'description', 'date', 'recurring', 'actions'];
     dataSource$: Observable<ITask[]>;
 
-    constructor(private taskService: TaskService) {
+    constructor(private taskService: TaskService, private dialog: MatDialog) {
         this.dataSource$ = this.taskService.tasks$; // Vincularse al observable del servicio
     }
 
-    deleteTask(index: number) {
-        this.taskService.deleteTask(index);
+    deleteTask(index: number): void {
+        const dialogRef = this.dialog.open(ModalComponent, {
+            width: '300px',
+            data: { message: '¿Estás seguro que deseas eliminar esta tarea?'}
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+            if(result) {
+                this.taskService.deleteTask(index);
+            }
+        });
     }
 
     toggleTask(index: number) {
